중복되는 부분 코드 정리
* UIConfigure - SlowlyRemoveIndicator 메소드 추가
* SetValueCurrentUser 메소드 추가
