+++ Modakyi/Like/LikeViewController.swift
@@ -31,13 +31,7 @@ class LikeViewController: UIViewController {
                 DispatchQueue.main.async {
                     self.labelView.isHidden = false
                     self.collectionview.reloadData()
-                    
-                    // 로딩뷰 천천히 없애기
-                    UIView.animate(withDuration: 0.5) {
-                        self.indicatorView.stopAnimating()
-                        self.indicatorView.alpha = 0
-                        self.collectionview.alpha = 1
-                    }
+                    SlowlyRemoveIndicator(self.indicatorView, self.collectionview)
                 }
                 return
             }
@@ -48,13 +42,7 @@ class LikeViewController: UIViewController {
             DispatchQueue.main.async {
                 self.labelView.isHidden = true
                 self.collectionview.reloadData()
-                
-                // 로딩뷰 천천히 없애기
-                UIView.animate(withDuration: 0.5) {
-                    self.indicatorView.stopAnimating()
-                    self.indicatorView.alpha = 0
-                    self.collectionview.alpha = 1
-                }
+                SlowlyRemoveIndicator(self.indicatorView, self.collectionview)
             }
         }
     }
